functions and files clean up
