+++ src/raytracing/camera.c
