+++ src/raytracing/shade.c
@@ -41,40 +41,35 @@ t_xyz	calc_shade(t_obj *obj, t_lit *lit, t_hit_point hit_obj, t_ray cam_ray)
 	return (pls_shade(obj, lit, dot_res, specular_ref));
 }
 
-// t_xyz	calc_shade(t_obj *obj, t_lit *lit, t_hit_point hit_obj, t_ray cam_ray)
-// {
-// 	t_xyz	col;
+t_xyz	pls_shade(t_obj *obj, t_lit *lit, double diff_ref, double spec_ref)
+{
+	t_xyz	dif_col;
+	t_xyz	spec_col;
+	t_xyz	ret_col;
+	t_xyz	lit_rgb;
 
-// 	(void)obj;
-// 	(void)lit;
-// 	(void)cam_ray;
-// 	col.x = (hit_obj.norm.x + 1.0) * 0.5 * 255.0;
-// 	col.y = (hit_obj.norm.y + 1.0) * 0.5 * 255.0;
-// 	col.z = (hit_obj.norm.z + 1.0) * 0.5 * 255.0;
-// 	return (col);
-// }
+	lit_rgb = lit->rgb;
+	lit_rgb = vec_div(lit_rgb, 255.0);
+	dif_col = vec_scale(vec_mul(obj->rgb, lit_rgb), diff_ref);
+	dif_col = vec_scale(dif_col, lit->t);
+	spec_col = vec_scale(vec_scale(lit_rgb, 255.0), spec_ref);
+	spec_col = vec_scale(spec_col, lit->t);
+	ret_col = vec_add(dif_col, spec_col);
+	return (ret_col);
+}
 
-// t_xyz	calc_shade(t_obj *obj, t_lit *lit, t_hit_point hit_obj, t_ray cam_ray)
-// {
-// 	t_xyz	col;
 
-// 	(void)lit;
-// 	(void)cam_ray;
-// 	col.x = 0.0;
-// 	col.y = 0.0;
-// 	col.z = 0.0;
-// 	if (obj->id == CY)
-// 	{
-// 		if (hit_obj.part == HIT_CY_SIDE)
-// 			col.y = 255.0;
-// 		else if (hit_obj.part == HIT_CY_CAP_TOP)
-// 			col.x = 255.0;
-// 		else if (hit_obj.part == HIT_CY_CAP_BOTTOM)
-// 			col.z = 255.0;
-// 		return (col);
-// 	}
-// 	col.x = 255.0;
-// 	col.y = 255.0;
-// 	col.z = 255.0;
-// 	return (col);
-// }
+int	set_amb_col(t_xyz *color, t_env *env)
+{
+	color->x = (env->amb_rgb.x / 255.0) * env->amb_t;
+	color->y = (env->amb_rgb.y / 255.0) * env->amb_t;
+	color->z = (env->amb_rgb.z / 255.0) * env->amb_t;
+	return (0);
+}
+
+void	pls_amb_color(t_obj *obj, t_env *env, t_xyz *col)
+{
+	col->x = obj->rgb.x * (env->amb_rgb.x / 255.0) * env->amb_t;
+	col->y = obj->rgb.y * (env->amb_rgb.y / 255.0) * env->amb_t;
+	col->z = obj->rgb.z * (env->amb_rgb.z / 255.0) * env->amb_t;
+}
