+++ src/ui/init_window.c
@@ -51,100 +51,3 @@ int	render_window(t_mlx_env *mlx, t_obj *obj, t_env *env)
 	render_scene(mlx, obj, env);
 	return (EXIT_SUCCESS);
 }
-
-t_obj	get_indexed_obj(int index, t_obj *obj)
-{
-	t_obj	*ret;
-	int		i;
-
-	if (!obj)
-		print_error_and_exit("get_indexed_obj", "obj is null");
-	if (index < 0)
-		print_error_and_exit("get_indexed_obj", "invalid index");
-	ret = obj;
-	i = 0;
-	while (ret && i < index)
-	{
-		ret = ret->next;
-		i++;
-	}
-	if (!ret)
-		print_error_and_exit("get_indexed_obj", "index out of range");
-	return (*ret);
-}
-
-void	reset_light_flags(t_env *env)
-{
-	t_lit	*l = env->lit;
-	while (l)
-	{
-		l->valid_flag = true;
-		l = l->next;
-	}
-}
-
-int	render_scene(t_mlx_env *mlx, t_obj *obj, t_env *env)
-{
-	int		x;
-	int		y;
-	t_xyz	color;
-	t_xyz	screen_vec;
-	t_ray	cam_ray;
-
-	y = -1;
-	while (++y < W_HEIGHT)
-	{
-		x = -1;
-		while (++x < W_WIDTH)
-		{
-			init_xyz(&color);
-			set_screen_vector(&screen_vec, x, y, env->cam_degree);
-			cam_ray.pos = env->cam_xyz;
-			cam_ray.dir = calc_cam_dir(screen_vec, env->cam_vector);
-			reset_light_flags(env);
-			check_light_pos(obj, env, cam_ray);
-			ray_tracing(obj, env, cam_ray, &color);
-			color_set_to_pixel \
-				(mlx->img, x, y, make_trgb(0, color.x, color.y, color.z));
-		}
-	}
-	mlx_put_image_to_window(mlx->mlx, mlx->window, mlx->img->img, 0, 0);
-	mlx_destroy_image(mlx->mlx, mlx->img->img);
-	return (EXIT_SUCCESS);
-}
-
-int	ray_tracing(t_obj *obj, t_env *env, t_ray cam_ray, t_xyz *color)
-{
-	t_hit_point	hit_obj;
-	t_lit		*tmp_lit;
-	t_obj		cpy_obj;
-	int			ret;
-
-	hit_obj.index = hit_nearest_obj(obj, &cam_ray, &hit_obj);
-	if (hit_obj.index < 0)
-		return (set_amb_col(color, env), 0);
-	cpy_obj = get_indexed_obj(hit_obj.index, obj);
-	fill_hit_obj(&cpy_obj, cam_ray, &hit_obj);
-	tmp_lit = env->lit;
-	pls_amb_color(&cpy_obj, env, color);
-	// while (tmp_lit && tmp_lit->valid_flag)
-	// {
-	// 	ret = calc_shadow(obj, tmp_lit, &hit_obj);
-	// 	if (ret == NOT_RENDERED_SHADOW)
-	// 		*color = vec_add(
-	// 			calc_shade(&cpy_obj, tmp_lit, hit_obj, cam_ray), *color);
-	// 	tmp_lit = tmp_lit->next;
-	// }
-	while (tmp_lit)
-	{
-		if (tmp_lit->valid_flag)
-		{
-			ret = calc_shadow(obj, tmp_lit, &hit_obj);
-			if (ret == NOT_RENDERED_SHADOW)
-				*color = vec_add(calc_shade(&cpy_obj, tmp_lit, hit_obj, cam_ray), *color);
-		}
-		tmp_lit = tmp_lit->next;
-	}
-	return (clamp_xyz(color, 0, 255), 1);
-	return (0);
-}
